+++ community/tests/unit_tests/comment_tests/persistent_tests.py
@@ -1,8 +1,6 @@
 import pytest
-from unittest.mock import Mock
 
 from community.model.access import *
-from .conftest import *
 
 
 class TestAccessFunctions:
