+++ community/tests/unit_tests/comment_tests/service_tests.py
@@ -1,8 +1,7 @@
 import pytest
 from unittest.mock import Mock
 
-from community.model.models import Post, Comment
-from .conftest import *
+from community.model.models import Comment
 from community.service.comment_service import *
 from community.domain.definition import PostFilesState
