+++ conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+pytest_plugins = [
+    "community.tests.fixtures.api_fixtures",
+    "community.tests.fixtures.mock_fixtures",
+    "community.tests.fixtures.instance_fixtures",
+    "community.tests.fixtures.data_fixtures",
+]
